Organize the networking classes in a better way
